my firt commit in develop , add change of the lenguaje to js for ts ,and add lint of tinkin
