--- lib/config.js
@@ -1,7 +1,0 @@
-export const config = {
-  dataset: process.env.NEXT_PUBLIC_SANITY_DATASET || "production",
-  projectId: process.env.NEXT_PUBLIC_SANITY_PROJECT_ID,
-  apiVersion: "2021-03-25",
-  useCdn: process.env.NODE_ENV === "production",
-  domain: process.env.DOMAIN,
-};
