--- pages/_app.js
@@ -1,7 +1,0 @@
-
-import Layout from '../components/layout'
-function MyApp({ Component, pageProps }) {
-  return <Layout><Component {...pageProps} /></Layout>
-}
-
-export default MyApp
