--- pages/api/twitch.js
@@ -1,4 +1,0 @@
-export default function handler(req, res) {
-    res.status(200).json({ name: 'Bienvenido a el stream de pamshojs' })
-  }
-
