--- pages/index.js
@@ -1,49 +1,0 @@
-import Head from "next/head";
-import Image from "next/image";
-import HeadHero from "../components/home/head";
-import Articles from "../components/home/articles";
-import {config} from "../lib/config"
-
-import { groq } from "next-sanity";
-
-import { getClient } from "../lib/sanity.server";
-
-const postQuery = groq`
-  *[_type == "post" ][0..4] {
-    _id,
-    title,
-  description,
-    mainImage,
-    body,
-    slug,
-    publishedAt
-    
-  }
-`;
-export default function Home({data,domain}) {
- 
-  return (
-    <div>
-      <Head>
-        <title>Pamsho.dev</title>
-        <meta name="description" content="Generated by create next app" />
-       
-      </Head>
-      <HeadHero />
-      <Articles data={data} domain={domain}/>
-    </div>
-  );
-}
-
-
-export async function getStaticProps({  preview = false }) {
-  const post = await getClient(preview).fetch(postQuery);
-
-  return {
-    props: {
-      preview,
-      data:post,
-      domain: config.domain,
-    },
-  };
-}
